Refactor worker script
* Move the changelog insertion out into a separate module as that took up most of the code.
* Move the initialisation of the resolvers with static configuration out of the per-message part of the worker so it occurs once on startup instead of initialising a new resolver per-message.
